+++ app/clients/views/client.py
@@ -1,6 +1,5 @@
-import json
-
 from django.http import Http404
+from django.core import validators
 
 from rest_framework.views import APIView
 from rest_framework.response import Response
@@ -10,11 +9,26 @@ from clients.models import Client
 from clients.serializers import ClientSerializer
 
 
-class ClientList(APIView):
+class ClientView(APIView):
+    def get_client_by_id(self, client_id):
+        try:
+            return Client.objects.get(id=client_id)
+        except (validators.ValidationError, Client.DoesNotExist):
+            raise Http404
+
+
     def get(self, request):
-        data = Client.objects.all()
-        serializer = ClientSerializer(data, many=True, context={'request': request})
-        return Response(serializer.data)
+        client_id = request.data.get('id', None) 
+        if client_id is None:
+            # get list of clients if no id provided
+            data = Client.objects.all()
+            serializer = ClientSerializer(data, many=True, context={'request': request})
+            return Response(serializer.data)
+        else:
+            # retrieve info of a client
+            client = self.get_client_by_id(client_id)
+            serializer = ClientSerializer(client)
+            return Response(serializer.data)
 
 
     def post(self, request):
@@ -26,22 +40,12 @@ class ClientList(APIView):
         return Response(serializer.data)
 
 
-class ClientDetail(APIView):
-    def get_object(self, client_pk):
-        try:
-            return Client.objects.get(pk=client_pk)
-        except Client.DoesNotExist:
-            raise Http404
+    def patch(self, request):
+        client_id = request.data.get('id', None) 
+        if client_id is None:
+            return Response(u'field "id" required in the url', status=status.HTTP_400_BAD_REQUEST)
 
-
-    def get(self, request, client_pk):
-        client = self.get_object(client_pk)
-        serializer = ClientSerializer(client)
-        return Response(serializer.data)
-
-
-    def patch(self, request, client_pk):
-        client = self.get_object(client_pk)
+        client = self.get_client_by_id(client_id)
         serializer = ClientSerializer(client, data=request.data, partial=True, context={'request': request})
         if not serializer.is_valid():
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
@@ -50,8 +54,12 @@ class ClientDetail(APIView):
         return Response(serializer.data)
 
 
-    def delete(self, request, client_pk):
-        client = self.get_object(client_pk)
+    def delete(self, request):
+        client_id = request.data.get('id', None) 
+        if client_id is None:
+            return Response(u'field "id" required in the url', status=status.HTTP_400_BAD_REQUEST)
+
+        client = self.get_client_by_id(client_id)
         client.delete()
         serializer = ClientSerializer(client)
         return Response(serializer.data)
